refactoring the ms reports
